+++ .gitignore
@@ -1,2 +1,2 @@
-/target
+/http_server/target
 /sql_server/target
