+++ sql_server/src/word.rs
